requirements.txt
Flask==3.1.0
Pillow==11.0.0
tensorflow==2.18.0
numpy==1.26.4
gunicorn==23.0.0
